requirements.txt
pygame==2.5.2
paho-mqtt
django-environ==0.11.2